refactor(model): group v2 events models in txn package, drop Txn prefix
